+++ core/models/resthooks.go
@@ -3,13 +3,11 @@ package models
 import (
 	"context"
 	"database/sql"
-	"time"
 
 	"github.com/jmoiron/sqlx"
 	"github.com/nyaruka/gocommon/dbutil"
 	"github.com/nyaruka/goflow/assets"
 	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
 )
 
 // ResthookID is our type for the database id of a resthook
@@ -35,8 +33,6 @@ func (r *Resthook) Subscribers() []string { return r.r.Subscribers }
 
 // loads the resthooks for the passed in org
 func loadResthooks(ctx context.Context, db *sql.DB, orgID OrgID) ([]assets.Resthook, error) {
-	start := time.Now()
-
 	rows, err := db.QueryContext(ctx, selectResthooksSQL, orgID)
 	if err != nil {
 		return nil, errors.Wrapf(err, "error querying resthooks for org: %d", orgID)
@@ -53,8 +49,6 @@ func loadResthooks(ctx context.Context, db *sql.DB, orgID OrgID) ([]assets.Resth
 
 		resthooks = append(resthooks, resthook)
 	}
-
-	logrus.WithField("elapsed", time.Since(start)).WithField("org_id", orgID).WithField("count", len(resthooks)).Debug("loaded resthooks")
 
 	return resthooks, nil
 }