+++ core/tasks/handler/contact_tasks.go
@@ -242,20 +242,28 @@ func HandleChannelEvent(ctx context.Context, rt *runtime.Runtime, eventType mode
 		}
 	}
 
+	var flowOptIn *flows.OptIn
+	if eventType == models.EventTypeOptIn || eventType == models.EventTypeOptOut {
+		optInID := models.OptInID(event.ExtraInt("optin_id"))
+		optIn := oa.OptInByID(optInID)
+		if optIn != nil {
+			flowOptIn = oa.SessionAssets().OptIns().Get(optIn.UUID())
+		}
+	}
+
 	// build our flow trigger
-	var flowTrigger flows.Trigger
+	tb := triggers.NewBuilder(oa.Env(), flow.Reference(), contact)
+	var trig flows.Trigger
 
 	if eventType == models.EventTypeIncomingCall {
 		urn := contacts[0].URNForID(event.URNID())
-		flowTrigger = triggers.NewBuilder(oa.Env(), flow.Reference(), contact).
-			Channel(channel.ChannelReference(), triggers.ChannelEventTypeIncomingCall).
-			WithCall(urn).
-			Build()
+		trig = tb.Channel(channel.ChannelReference(), triggers.ChannelEventTypeIncomingCall).WithCall(urn).Build()
+	} else if eventType == models.EventTypeOptIn && flowOptIn != nil {
+		trig = tb.OptIn(flowOptIn, triggers.OptInEventTypeStarted).Build()
+	} else if eventType == models.EventTypeOptOut && flowOptIn != nil {
+		trig = tb.OptIn(flowOptIn, triggers.OptInEventTypeStopped).Build()
 	} else {
-		flowTrigger = triggers.NewBuilder(oa.Env(), flow.Reference(), contact).
-			Channel(channel.ChannelReference(), triggers.ChannelEventType(eventType)).
-			WithParams(params).
-			Build()
+		trig = tb.Channel(channel.ChannelReference(), triggers.ChannelEventType(eventType)).WithParams(params).Build()
 	}
 
 	// if we have a channel connection we set the connection on the session before our event hooks fire
@@ -270,7 +278,7 @@ func HandleChannelEvent(ctx context.Context, rt *runtime.Runtime, eventType mode
 		}
 	}
 
-	sessions, err := runner.StartFlowForContacts(ctx, rt, oa, flow, []*models.Contact{modelContact}, []flows.Trigger{flowTrigger}, hook, true)
+	sessions, err := runner.StartFlowForContacts(ctx, rt, oa, flow, []*models.Contact{modelContact}, []flows.Trigger{trig}, hook, true)
 	if err != nil {
 		return nil, errors.Wrapf(err, "error starting flow for contact")
 	}